Trap NPE in JobStoreSupport#updateMisfiredTrigger
Implements a workaround for QTZ-450.

This workaround simply replaces an NPE by returning a value
of false for the misfire -- as if no misfire had happened.

There are two opportunities for an NPE in that
line of code, and I didn't know which one caused it, so I trap both.
I log the NPE but return false.

This change does not address the real problem, that there is a null
reference.

https://jira.terracotta.org/jira/browse/QTZ-450